[REFACTOR] 문자열 분리용 객체 코드 정리 (getter 제거, 테스트 이동, 상수정의)
